+++ CHANGELOG.md
@@ -57,6 +57,9 @@ In order to build the DUNE core modules you need at least the following software
 - Add the utility `IsCompileTimeConstant` to check for integral constants and anything with
   the same interface.
 
+- Add dedicated includes `dune/common/metis.hh` for METIS and `dune/common/parallel/parmetis.hh`
+  for ParMETIS to be used instead of the direct includes `metis.h` and `parmetis.h`.
+
 ## Build System
 
 - Documentation files in `doc/buildsystem/${ModuleName}.rst` are now only copied.